{#2,#3,#4,#6} Refactoring and documentation
