+++ app/src/main/res/layout/fragment_initial_selection.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="de.teamawesome.awesomeplayer.InitialSelectionFragment">
+    tools:context="de.teamawesome.awesomeplayer.fragments.InitialSelectionFragment">
 
     <Button
         android:id="@+id/Button_All"
